moved files to controllers directory. refactored test scripts, error.php.
