+++ php/logs/error_log.txt
@@ -42,3 +42,18 @@ newchat does not exist
  does not exist
  does not exist
  does not exist
+ does not exist
+ does not exist
+ does not exist
+ does not exist
+ does not exist
+ does not exist
+ does not exist
+ does not exist
+ does not exist
+ does not exist
+ does not exist
+ does not exist
+ does not exist
+ does not exist
+ does not exist
