--- php/pages/logout.php
@@ -1,5 +1,0 @@
-<?php
-    session_start();
-    session_destroy();
-    header('Location: ./login.php');
-?>
