--- php/tests/error_log.txt
@@ -1,22 +1,0 @@
-Error: INSERT INTO friends('from_id', 'to_id', 'pending') VALUES(, , false) 
-You have an error in your SQL syntax; check the manual that corresponds to your MySQL server version for the right syntax to use near ''from_id', 'to_id', 'pending') VALUES(, , false)' at line 1Error: INSERT INTO friends('from_id', 'to_id', 'pending') VALUES(, , false) 
-You have an error in your SQL syntax; check the manual that corresponds to your MySQL server version for the right syntax to use near ''from_id', 'to_id', 'pending') VALUES(, , false)' at line 1Error: INSERT INTO friends('from_id', 'to_id', 'pending') VALUES(, , false) 
-You have an error in your SQL syntax; check the manual that corresponds to your MySQL server version for the right syntax to use near ''from_id', 'to_id', 'pending') VALUES(, , false)' at line 1Error: INSERT INTO friends('from_id', 'to_id', 'pending') VALUES(123, 124, false) 
-You have an error in your SQL syntax; check the manual that corresponds to your MySQL server version for the right syntax to use near ''from_id', 'to_id', 'pending') VALUES(123, 124, false)' at line 1Error: INSERT INTO friends('from_id', 'to_id', 'pending') VALUES(123, 124, false) 
-You have an error in your SQL syntax; check the manual that corresponds to your MySQL server version for the right syntax to use near ''from_id', 'to_id', 'pending') VALUES(123, 124, false)' at line 1Error: INSERT INTO friends('from_id', 'to_id', 'pending') VALUES(131, 132, false) 
-You have an error in your SQL syntax; check the manual that corresponds to your MySQL server version for the right syntax to use near ''from_id', 'to_id', 'pending') VALUES(131, 132, false)' at line 1Error: INSERT INTO friends('from_id', 'to_id', 'pending') VALUES(135, 136, false) 
-You have an error in your SQL syntax; check the manual that corresponds to your MySQL server version for the right syntax to use near ''from_id', 'to_id', 'pending') VALUES(135, 136, false)' at line 1Error: DELETE FROM user WHERE name = 'Sender User' AND password = '' 
-Cannot delete or update a parent row: a foreign key constraint fails (`mio_db`.`friends`, CONSTRAINT `fk_friends_user_from` FOREIGN KEY (`from_id`) REFERENCES `user` (`id`))Error: SELECT to_id FROM friends WHERE from_id =  
-You have an error in your SQL syntax; check the manual that corresponds to your MySQL server version for the right syntax to use near '' at line 1Error: DELETE FROM user WHERE name = 'Sender User' AND password = '' 
-Cannot delete or update a parent row: a foreign key constraint fails (`mio_db`.`friends`, CONSTRAINT `fk_friends_user_from` FOREIGN KEY (`from_id`) REFERENCES `user` (`id`))Error: SELECT to_id FROM friends WHERE from_id =  
-You have an error in your SQL syntax; check the manual that corresponds to your MySQL server version for the right syntax to use near '' at line 1Error: DELETE FROM user WHERE name = 'Sender User' AND password = '' 
-Cannot delete or update a parent row: a foreign key constraint fails (`mio_db`.`friends`, CONSTRAINT `fk_friends_user_from` FOREIGN KEY (`from_id`) REFERENCES `user` (`id`))Error: SELECT to_id FROM friends WHERE from_id =  
-You have an error in your SQL syntax; check the manual that corresponds to your MySQL server version for the right syntax to use near '' at line 1Error: DELETE FROM user WHERE name = 'Sender User' AND password = '' 
-Cannot delete or update a parent row: a foreign key constraint fails (`mio_db`.`friends`, CONSTRAINT `fk_friends_user_from` FOREIGN KEY (`from_id`) REFERENCES `user` (`id`))Error: SELECT to_id FROM friends WHERE from_id =  
-You have an error in your SQL syntax; check the manual that corresponds to your MySQL server version for the right syntax to use near '' at line 1Error: DELETE FROM user WHERE name = 'Sender User' AND password = '' 
-Cannot delete or update a parent row: a foreign key constraint fails (`mio_db`.`friends`, CONSTRAINT `fk_friends_user_from` FOREIGN KEY (`from_id`) REFERENCES `user` (`id`))Error: SELECT to_id FROM friends WHERE from_id =  
-You have an error in your SQL syntax; check the manual that corresponds to your MySQL server version for the right syntax to use near '' at line 1Error: DELETE FROM user WHERE name = 'Sender User' AND password = '' 
-Cannot delete or update a parent row: a foreign key constraint fails (`mio_db`.`friends`, CONSTRAINT `fk_friends_user_from` FOREIGN KEY (`from_id`) REFERENCES `user` (`id`))Error: SELECT to_id FROM friends WHERE from_id =  
-You have an error in your SQL syntax; check the manual that corresponds to your MySQL server version for the right syntax to use near '' at line 1Error: DELETE FROM user WHERE name = 'Sender User' AND password = '' 
-Cannot delete or update a parent row: a foreign key constraint fails (`mio_db`.`friends`, CONSTRAINT `fk_friends_user_from` FOREIGN KEY (`from_id`) REFERENCES `user` (`id`))Error: SELECT to_id FROM friends WHERE from_id =  
-You have an error in your SQL syntax; check the manual that corresponds to your MySQL server version for the right syntax to use near '' at line 1
