+++ php/views/logout.php
@@ -1,0 +1,6 @@
+<?php
+    require_once ("../router/redirect.php");
+    $_SESSION = array();
+    session_destroy();
+    redirect("login.php");
+    die();
