+++ scripts/socket.js
@@ -1,4 +1,5 @@
 var websocket;
+let websocketUrl = "ws://localhost:8080/php/chat_server/socketServer.php";
 //websocket readyState constants
 let CONNECTING = 0;
 let OPEN = 1;
@@ -7,7 +8,7 @@ let CLOSED = 3;
 //end websocket readyState constants
 
 function createSocket() {
-    let socket = new WebSocket("ws://localhost:8080/php/chat_server/socketServer.php");
+    let socket = new WebSocket(websocketUrl);
     initializeSocketEventHandlers(socket);
     return socket;
 }
